requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
azure-storage-blob==12.19.0
azure-storage-queue==12.9.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
pydantic-settings==2.2.0
pytest==8.3.0
httpx==0.27.0
